better document API routes and split file
